requirements.txt
asgiref==3.7.2
blinker==1.6.3
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
coverage==7.3.2
Django==4.2.6
django-environ==0.11.2
django-filter==23.3
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
environ==1.0
exceptiongroup==1.1.3
flake8==6.1.0
flake8-functions==0.0.8
flake8-html==0.4.3
gevent==23.9.1
geventhttpclient==2.0.11
greenlet==3.0.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mr-proper==0.0.7
msgpack==1.0.7
packaging==23.2
phonenumbers==8.13.23
pluggy==1.3.0
psutil==5.9.6
psycopg2==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
Pygments==2.16.1
PyJWT==2.8.0
pytest==7.4.2
pytest-cov==4.1.0
pytest-django==4.6.0
pytz==2023.3.post1
pywin32==306; platform_system=="Windows"
pyzmq==25.1.1
requests==2.31.0
roundrobin==0.0.4
sentry-sdk==1.39.1
six==1.16.0
sqlparse==0.4.4
stdlib-list==0.9.0
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
Werkzeug==3.0.0
zope.event==5.0
zope.interface==6.1
